+++ tests/urls.py
@@ -18,6 +18,7 @@ urlpatterns = [
     path("new_user/", views.new_user),
     path("execute_sql/", views.execute_sql),
     path("cached_view/", views.cached_view),
+    path("json_view/", views.json_view),
     path("redirect/", views.redirect_view),
     path("__debug__/", include(debug_toolbar.urls)),
 ]