+++ docs/changes.rst
@@ -23,6 +23,7 @@ Change log
 * Added support for ``cache.touch()`` when using django-debug-toolbar.
 * Eliminated more inline CSS.
 * Updated ``tox.ini`` and ``Makefile`` to use isort>=5.
+* Increased RESULTS_CACHE_SIZE to 25 to better support AJAX requests.
 
 **Backwards incompatible changes**
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -32,6 +33,8 @@ Change log
   :attr:`Panel.scripts <debug_toolbar.panels.Panel.scripts>` property.
 * Removed support for end of life Django 1.11. The minimum supported Django is
   now 2.2.
+
+* Added HistoryPanel including support for AJAX requests.
 
 
 2.2 (2020-01-31)